requirements.txt
pybloom-live==4.0.0
onnxruntime==1.15.1
ctranslate2==3.19.0
scrubadub==2.0.1
fasttext==0.9.2
re
datasets
json
pandas
numpy
tqdm.auto
typing
embedding
threading
huggingface_hub
networkx
sklearn.cluster
transformers
colorama
tiktoken